+++ CHANGELOG.md
@@ -1,3 +1,46 @@
+## [0.10.3](https://github.com/adobe/helix-shared/compare/v0.10.2...v0.10.3) (2019-04-11)
+
+
+### Bug Fixes
+
+* **yaml:** Avoid deprecation warnings from yaml ([#89](https://github.com/adobe/helix-shared/issues/89)) ([1374cfe](https://github.com/adobe/helix-shared/commit/1374cfe)), closes [#88](https://github.com/adobe/helix-shared/issues/88)
+
+## [0.10.2](https://github.com/adobe/helix-shared/compare/v0.10.1...v0.10.2) (2019-04-10)
+
+
+### Bug Fixes
+
+* **package:** update yaml to version 1.5.0 ([bf42d94](https://github.com/adobe/helix-shared/commit/bf42d94))
+
+## [0.10.1](https://github.com/adobe/helix-shared/compare/v0.10.0...v0.10.1) (2019-04-10)
+
+
+### Bug Fixes
+
+* **config:** Ensure default ref is added to string giturls ([#87](https://github.com/adobe/helix-shared/issues/87)) ([30d9cc3](https://github.com/adobe/helix-shared/commit/30d9cc3)), closes [#86](https://github.com/adobe/helix-shared/issues/86)
+
+# [0.10.0](https://github.com/adobe/helix-shared/compare/v0.9.0...v0.10.0) (2019-03-27)
+
+
+### Features
+
+* Helpers for implementing improved frontmatter in helix-pipe ([d5544d5](https://github.com/adobe/helix-shared/commit/d5544d5))
+* Improve sequence.js dealing with null values ([f79fa59](https://github.com/adobe/helix-shared/commit/f79fa59))
+
+# [0.9.0](https://github.com/adobe/helix-shared/compare/v0.8.4...v0.9.0) (2019-03-27)
+
+
+### Features
+
+* **configuration:** Strains can be and should be specified as an ordered list ([38b5e9d](https://github.com/adobe/helix-shared/commit/38b5e9d)), closes [#71](https://github.com/adobe/helix-shared/issues/71)
+
+## [0.8.4](https://github.com/adobe/helix-shared/compare/v0.8.3...v0.8.4) (2019-03-23)
+
+
+### Bug Fixes
+
+* **config:** Add missing performance threshold properties ([a90a489](https://github.com/adobe/helix-shared/commit/a90a489)), closes [#65](https://github.com/adobe/helix-shared/issues/65)
+
 ## [0.8.3](https://github.com/adobe/helix-shared/compare/v0.8.2...v0.8.3) (2019-03-22)